refactor: extract request-log account filter helpers
Move owned-key collapse and selector expansion out of request_log_query.go
so the file stays under the 1200-line backend structure gate.
